+++ core/src/main/java/com/ensody/reactivestate/Derived.kt
@@ -44,7 +44,7 @@ public fun <T> CoroutineScope.derived(
  * Creates a [StateFlow] that computes its value based on other [StateFlow]s via a suspendable [coAutoRun] block.
  *
  * @param initial The initial value (until the first computation finishes).
- * @param flowTransformer How changes should be collected. Defaults to `{ mapLatest { } }`.
+ * @param flowTransformer How changes should be executed/collected. Defaults to `{ conflatedWorker() }`.
  * @param dispatcher The [CoroutineDispatcher] to use. Defaults to `dispatchers.default`.
  * @param withLoading Whether loading state may be tracked for the (re-)computation. Defaults to `true`.
  * @param observer The callback which is used to track the observables.
@@ -52,7 +52,7 @@ public fun <T> CoroutineScope.derived(
 public fun <T> CoroutineLauncher.derived(
     initial: T,
     started: SharingStarted,
-    flowTransformer: AutoRunFlowTransformer = { mapLatest { } },
+    flowTransformer: AutoRunFlowTransformer = defaultAutoRunFlowTransformer,
     dispatcher: CoroutineDispatcher = dispatchers.default,
     withLoading: Boolean = true,
     observer: CoAutoRunCallback<T>,
@@ -82,7 +82,7 @@ public fun <T> CoroutineLauncher.derived(
  *
  * @param initial The initial value (until the first computation finishes).
  * @param launcher The [CoroutineLauncher] to use.
- * @param flowTransformer How changes should be collected. Defaults to `{ mapLatest { } }`.
+ * @param flowTransformer How changes should be executed/collected. Defaults to `{ conflatedWorker() }`.
  * @param dispatcher The [CoroutineDispatcher] to use. Defaults to `dispatchers.default`.
  * @param withLoading Whether loading state may be tracked for the (re-)computation. Defaults to `true`.
  * @param observer The callback which is used to track the observables.
@@ -91,7 +91,7 @@ public fun <T> CoroutineScope.derived(
     initial: T,
     started: SharingStarted,
     launcher: CoroutineLauncher = SimpleCoroutineLauncher(this),
-    flowTransformer: AutoRunFlowTransformer = { mapLatest { } },
+    flowTransformer: AutoRunFlowTransformer = defaultAutoRunFlowTransformer,
     dispatcher: CoroutineDispatcher = dispatchers.default,
     withLoading: Boolean = true,
     observer: CoAutoRunCallback<T>,